adjustments to sql-schema.sql

+++ src/main/java/com/qa/ims/controller/ItemController.java
@@ -46,7 +46,7 @@ public class ItemController implements CrudController<Item> {
         LOGGER.info("Please enter an item name");
         String itemName = utils.getString();
         LOGGER.info("Please enter the items price");
-        Float price = Float.parseFloat(utils.getString());
+        Double price = Double.parseDouble(utils.getString());
         Item item = itemDAO.create(new Item(itemName, price));
         LOGGER.info("Item created");
         return item;
@@ -62,7 +62,7 @@ public class ItemController implements CrudController<Item> {
         LOGGER.info("Please enter an item name");
         String itemName = utils.getString();
         LOGGER.info("Please enter a price");
-        Float price = Float.parseFloat(utils.getString());
+        Double price = Double.parseDouble(utils.getString());
         Item item = itemDAO.update(new Item(id, itemName, price));
         LOGGER.info("Item Updated");
         return item;
